backend/requirements.txt
bleach>=3.1.0
django>=2.2
djangorestframework>=3.10.3
djangorestframework-simplejwt>=4.3.0
django-cleanup>=4.0.0
django-enumfield>=1.5
django-filter>=2.2.0
gunicorn>=19.9.0
pillow>=6.2.1
psycopg2>=2.8.2
requests>=2.22.0
